Add mark as favorite action on media action modal

+++ core/domain/src/main/kotlin/com/divinelink/core/domain/MarkAsFavoriteUseCase.kt
@@ -6,18 +6,17 @@ import com.divinelink.core.commons.domain.data
 import com.divinelink.core.data.media.repository.MediaRepository
 import com.divinelink.core.model.media.MediaItem
 
-@Suppress("ThrowingExceptionsWithoutMessageOrCause")
 open class MarkAsFavoriteUseCase(
   private val repository: MediaRepository,
   dispatcher: DispatcherProvider,
-) : UseCase<MediaItem.Media, Unit>(dispatcher.io) {
-  override suspend fun execute(parameters: MediaItem.Media) {
+) : UseCase<MediaItem.Media, Boolean>(dispatcher.default) {
+  override suspend fun execute(parameters: MediaItem.Media): Boolean {
     val isFavorite = repository.checkIfMediaIsFavorite(
       id = parameters.id,
       mediaType = parameters.mediaType,
     ).data
 
-    val result = if (isFavorite) {
+    if (isFavorite) {
       repository.removeFavoriteMedia(
         id = parameters.id,
         mediaType = parameters.mediaType,
@@ -26,15 +25,6 @@ open class MarkAsFavoriteUseCase(
       repository.insertFavoriteMedia(parameters)
     }
 
-    return result.data
-//    if (result.isSuccess) {
-//      result.data
-//    } else if (result.isFailure) {
-//      throw result.exceptionOrNull()!!
-//    }
-//      when (result) {
-//      isSuccess -> result.data
-//      is result -> throw result.exception
-//    }
+    return !isFavorite
   }
 }
